backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
python-dotenv==1.0.0
torch==2.1.0
transformers==4.35.0
qdrant-client==1.6.0
langchain==0.0.350
pypdf==3.17.1
ollama==0.1.6
gunicorn==21.2.0
accelerate==0.25.0 